+++ packages/rpc_dart_transports/lib/rpc_dart_transports.dart
@@ -5,6 +5,9 @@
 /// Транспорты для RPC Dart
 library;
 
+export 'package:isolate_manager/isolate_manager.dart'
+    show isolateManagerCustomWorker;
+
 // Экспорт транспортов
 export 'src/adapters/secure/_index.dart';
 export 'src/server/_index.dart';